Change repository id type

+++ src/main/java/com/uah/f1backend/repository/TeamModelRepository.java
@@ -4,7 +4,7 @@ import com.uah.f1backend.model.TeamModel;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeamModelRepository extends JpaRepository<TeamModel, Long> {
+public interface TeamModelRepository extends JpaRepository<TeamModel, Integer> {
     // Find a team in db matching the given name
     Optional<TeamModel> findByName(String name);
 }
